vectrill::connectors

Trait Connector

pub trait Connector: Send + Sync {
    // Required methods
    fn next_batch<'life0, 'async_trait>(
        &'life0 mut self,
    ) -> Pin<Box<dyn Future<Output = Option<Result<RecordBatch>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn schema(&self) -> SchemaRef;
    fn name(&self) -> &str;
}

Trait for all data source connectors

Required Methods

fn next_batch<'life0, 'async_trait>( &'life0 mut self, ) -> Pin<Box<dyn Future<Output = Option<Result<RecordBatch>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get the next batch of data from the connector Returns None when the connector is exhausted

fn schema(&self) -> SchemaRef

Get the schema of the data produced by this connector

fn name(&self) -> &str

Get the name of this connector