vectrill::memory::buffer_pool

Struct BufferPool

pub struct BufferPool {
    pools: Mutex<HashMap<DataType, Vec<ArrayRef>>>,
    max_size_per_pool: usize,
}

Buffer pool for reusing Arrow arrays

Fields
pools: Mutex<HashMap<DataType, Vec<ArrayRef>>>

Pools keyed by data type

max_size_per_pool: usize

Maximum size per pool

Implementations

impl BufferPool

pub fn new(max_size_per_pool: usize) -> Self

Create a new buffer pool

pub fn get_array(&self, dtype: &DataType, capacity: usize) -> ArrayRef

Get an array from the pool or create a new one

pub fn return_array(&self, array: ArrayRef)

Return an array to the pool

fn create_array(&self, dtype: &DataType, capacity: usize) -> ArrayRef

Create a new array of the given type and capacity

pub fn clear(&self)

Clear all pools

pub fn stats(&self) -> PoolStats

Get statistics about the pool